patch: 2.1.0 - template-usge (#138)
* fix: Added templateId as label to create/update/delete student/group documents to also have metrics of which templates are in use

* fix: Expose info gauges of the template id-name mapping (to have human readable info in Grafana)

* chore: Moved metric constants to shared metric constant file

* fix: Use contentType from the register itself to explicitly set utf8

+++ src/lib/server/db/mongo/document-content-templates-db-client.ts
@@ -1,7 +1,16 @@
-import { type Collection, type Db, ObjectId } from "mongodb"
+import { type Collection, type Db, ObjectId, type WithId } from "mongodb"
 import type { IDocumentContentTemplatesDbClient } from "$lib/types/db/db-client"
 import type { AvailableForDocumentType, DocumentContentTemplate, MetricCount, MetricLabel, NewDocumentContentTemplate } from "$lib/types/db/shared-types"
-import { incrementCount, metricResultFailure, metricResultName, metricResultSuccessful } from "../../metrics/handle-metrics"
+import {
+  metricsDocumentTemplateIdLabelName,
+  metricsDocumentTemplateInfoDescription,
+  metricsDocumentTemplateInfoName,
+  metricsDocumentTemplateNameLabelName,
+  metricsResultFailure,
+  metricsResultName,
+  metricsResultSuccessful
+} from "$lib/utils/metric-constants.js"
+import { createInfoGauges, incrementCount, removeInfoGauge } from "../../metrics/handle-metrics"
 
 export class DocumentContentTemplatesDbClient implements IDocumentContentTemplatesDbClient {
   private documentContentTemplatesCollection: Collection<NewDocumentContentTemplate>
@@ -62,21 +71,38 @@ export class DocumentContentTemplatesDbClient implements IDocumentContentTemplat
     if (!result.insertedId) {
       incrementCount({
         ...metricBody,
-        labels: [...labels, [metricResultName, metricResultFailure]]
+        labels: [...labels, [metricsResultName, metricsResultFailure]]
       })
 
       throw new Error("Failed to create document template")
     }
 
+    createInfoGauges(metricsDocumentTemplateInfoName, [
+      {
+        name: template.name, // NOTE: This isn't actually used, but required in type
+        description: metricsDocumentTemplateInfoDescription,
+        value: 1,
+        labels: [
+          [metricsDocumentTemplateIdLabelName, result.insertedId.toString()],
+          [metricsDocumentTemplateNameLabelName, template.name]
+        ]
+      }
+    ])
+
     incrementCount({
       ...metricBody,
-      labels: [...labels, [metricResultName, metricResultSuccessful]]
+      labels: [...labels, [metricsResultName, metricsResultSuccessful]]
     })
 
     return result.insertedId.toString()
   }
 
   async updateDocumentContentTemplate(templateId: string, template: NewDocumentContentTemplate): Promise<string> {
+    const currentTemplate: WithId<NewDocumentContentTemplate> | null = await this.documentContentTemplatesCollection.findOne({ _id: new ObjectId(templateId) })
+    if (!currentTemplate) {
+      throw new Error("Current document template not found")
+    }
+
     const result = await this.documentContentTemplatesCollection.updateOne({ _id: new ObjectId(templateId) }, { $set: { ...template } })
 
     const metricBody: MetricCount = {
@@ -96,21 +122,43 @@ export class DocumentContentTemplatesDbClient implements IDocumentContentTemplat
     if (result.modifiedCount === 0) {
       incrementCount({
         ...metricBody,
-        labels: [...labels, [metricResultName, metricResultFailure]]
+        labels: [...labels, [metricsResultName, metricsResultFailure]]
       })
 
       throw new Error("Failed to update document content template")
     }
 
+    removeInfoGauge(metricsDocumentTemplateInfoName, [
+      [metricsDocumentTemplateIdLabelName, templateId],
+      [metricsDocumentTemplateNameLabelName, currentTemplate.name]
+    ])
+
+    createInfoGauges(metricsDocumentTemplateInfoName, [
+      {
+        name: template.name, // NOTE: This isn't actually used, but required in type
+        description: metricsDocumentTemplateInfoDescription,
+        value: 1,
+        labels: [
+          [metricsDocumentTemplateIdLabelName, templateId],
+          [metricsDocumentTemplateNameLabelName, template.name]
+        ]
+      }
+    ])
+
     incrementCount({
       ...metricBody,
-      labels: [...labels, [metricResultName, metricResultSuccessful]]
+      labels: [...labels, [metricsResultName, metricsResultSuccessful]]
     })
 
     return templateId
   }
 
   async deleteDocumentContentTemplate(templateId: string): Promise<void> {
+    const currentTemplate: WithId<NewDocumentContentTemplate> | null = await this.documentContentTemplatesCollection.findOne({ _id: new ObjectId(templateId) })
+    if (!currentTemplate) {
+      throw new Error("Document template not found")
+    }
+
     const result = await this.documentContentTemplatesCollection.deleteOne({ _id: new ObjectId(templateId) })
 
     const metricBody: MetricCount = {
@@ -121,15 +169,20 @@ export class DocumentContentTemplatesDbClient implements IDocumentContentTemplat
     if (result.deletedCount === 0) {
       incrementCount({
         ...metricBody,
-        labels: [[metricResultName, metricResultFailure]]
+        labels: [[metricsResultName, metricsResultFailure]]
       })
 
       throw new Error("Failed to delete document content template")
     }
 
+    removeInfoGauge(metricsDocumentTemplateInfoName, [
+      [metricsDocumentTemplateIdLabelName, templateId],
+      [metricsDocumentTemplateNameLabelName, currentTemplate.name]
+    ])
+
     incrementCount({
       ...metricBody,
-      labels: [[metricResultName, metricResultSuccessful]]
+      labels: [[metricsResultName, metricsResultSuccessful]]
     })
   }
 }
